requirements.txt
tensorflow-gpu
matplotlib
h5py
ipykernel
scipy
image
